requirements.txt
pandas
numpy
tqdm
scikit-learn
torch
sentence-transformers
faiss-cpu
rank_bm25
fastapi
uvicorn
pydantic
openai
